Add tests to behaviours.

+++ src/eru.Application/Common/Interfaces/IStopwatch.cs
@@ -4,6 +4,6 @@
     {
         void Start();
         void Stop();
-        int ElapsedMilliseconds { get; }
+        long ElapsedMilliseconds { get; }
     }
 }
